Tue, Apr 12, 2022, 3:45 AM +00:00

+++ readability/readability.c
@@ -21,7 +21,7 @@ int main(void)
     float l = (count_letters(input) / count_words(input)) * 100.0;
 
     //calc average no of sentences per 100 word
-    float w = ((float) count_sentences(input) / count_words(input)) * 100.0;
+    float w = (count_sentences(input) / count_words(input)) * 100.0;
 
     //index = 0.0588 * L - 0.296 * S - 15.8
     float index = round((0.0588 * l) - (0.296 * w) - 15.8);
